InfoBarTimeshift: allow pause/unpause on streams
When a stream is entered on bouquets pause/unpause doesn't work because
pause button tries to create timeshift.

Now when time shift fails and service is streamed it will try to pause/unpause
the stream, only when type is not DVB.

Based on 507b9bf5895717408ee5cd3d504e385ca392612a with extra check for DVB, so
that will work for types different than DVB.

It was reverted on 9a465bba5a13f683520f8da52b12ea411b44feb0 because it had
side effects that a started timeshift could not be paused anymore.