epgcache: Reduce "FixOverlapping" margin from 300 to 60 seconds
The FixOverlapping method appears to remove events with a 5 minute (300 second)
margin. I suspect this is responsible for double EPG entries when a program has
been moved only a few minutes in the schedule (old event does not get erased in
that case).

Change the margin to just 1 minute (60 seconds) to see if this fixes things. It
makes sense anyway, why would an event start up to 5 minutes before the current
one ends?